code reafactoring

+++ src/services/contactsServices/deleteContact.js
@@ -1,8 +1,8 @@
 const { Contact } = require('../../models');
 const { HttpError } = require('../../utils');
 
-const deleteContact = async contactId => {
-  const contact = await Contact.findOneAndRemove({ _id: contactId });
+const deleteContact = async (userId, contactId) => {
+  const contact = await Contact.findOneAndRemove({ _id: contactId, owner: userId });
   if (!contact) {
     throw new HttpError(404, `Contact with id: ${contactId} not found`);
   }
